Move parsing services into services directory
