+++ tests/Service/Parse/ControlSymbolTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Jstewmc\Rtf\Parser;
+namespace Jstewmc\Rtf\Service\Parse;
 
 use Jstewmc\Rtf\{
     Token\Control\Symbol as Token,
